Rename javascript files as jsx files
